+++ packages/applications/ssr/src/components/molecules/CopyAndPaste.tsx
@@ -1,0 +1,36 @@
+'use client';
+
+import Button from '@codegouvfr/react-dsfr/Button';
+import { useState } from 'react';
+
+type Props = {
+  textToCopy: string;
+  timeoutInMs?: number;
+};
+export const CopyAndPaste = ({ textToCopy, timeoutInMs = 2000 }: Props) => {
+  const [hasCopied, setHasCopied] = useState<boolean>(false);
+
+  const copyLink = () => {
+    navigator.clipboard.writeText(textToCopy);
+    setHasCopied(true);
+    setTimeout(() => {
+      setHasCopied(false);
+    }, timeoutInMs);
+  };
+
+  return (
+    <div className="flex flex-row gap-3 items-center">
+      <span className="italic">{textToCopy}</span>{' '}
+      <Button
+        iconId="ri-clipboard-line"
+        aria-label="copier-coller"
+        priority="secondary"
+        onClick={copyLink}
+        style={{ marginTop: 0 }}
+        size="small"
+      >
+        {hasCopied ? 'Copié !' : 'Copier'}
+      </Button>
+    </div>
+  );
+};
